* ms/correct-formating-in-exit:
  Correct formating in exit error messages

OTP-10148

+++ erts/emulator/drivers/unix/unix_efile.c
@@ -107,8 +107,8 @@ static void *ef_safe_alloc(Uint s)
 {
     void *p = EF_ALLOC(s);
     if (!p) erl_exit(1,
-		     "unix efile drv: Can't allocate %d bytes of memory\n",
-		     s);
+		     "unix efile drv: Can't allocate %lu bytes of memory\n",
+		     (unsigned long)s);
     return p;
 }
 
@@ -118,8 +118,8 @@ static void *ef_safe_realloc(void *op, Uint s)
 {
     void *p = EF_REALLOC(op, s);
     if (!p) erl_exit(1,
-		     "unix efile drv: Can't reallocate %d bytes of memory\n",
-		     s);
+		     "unix efile drv: Can't reallocate %lu bytes of memory\n",
+		     (unsigned long)s);
     return p;
 }
